[DEV-13004] Fix CalloutNode attribute normalization
