+++ packages/slate-editor/src/extensions/callout/lib/index.ts
@@ -1,5 +1,5 @@
 export { createCallout } from './createCallout';
 export { insertCallout } from './insertCallout';
-export { normalizeRedundantAttributes } from './normalizeRedundantAttributes';
+export { normalizeAttributes } from './normalizeAttributes';
 export { parseCalloutElement } from './parseCalloutElement';
 export { updateCallout } from './updateCallout';
